+++ build-scripts/traceur/TraceurEs6ify.ts
@@ -2,10 +2,11 @@
 import gutil = require("gulp-util");
 import vinylSourceStream = require("vinyl-source-stream");
 import watchify = require("watchify");
+import PathUtil = require("../PathUtil");
 import BrowserifyHelper = require("../BrowserifyHelper");
 import Es6ifyLike = require("./Es6ifyLike");
 
-var shortName = BrowserifyHelper.toShortFileName;
+var shortName = PathUtil.toShortFileName;
 
 
 module TraceurEs6ify {
@@ -13,7 +14,10 @@ module TraceurEs6ify {
     export function compileScripts(debug: boolean, verboseCompileInfo: boolean, paths: AppPaths) {
         var { dstDir, dstFile, entryFile } = paths;
 
-        var bundlerOpts = BrowserifyHelper.createOptions(Object.assign({ debug }, paths), [watchify]);
+        var bfyOpts: Browserify.Options & BrowserPack.Options = {
+            debug,
+        };
+        var bundlerOpts = BrowserifyHelper.createOptions(Object.assign(bfyOpts, paths), [watchify]);
         var bundler = BrowserifyHelper.create(bundlerOpts);
 
         Es6ifyLike.traceurOverrides.global = true;