+++ gulpfile.ts
@@ -2,6 +2,7 @@
 import gutil = require("gulp-util");
 import BabelBabelify = require("./build-scripts/babel/BabelBabelify");
 import TraceurEs6ify = require("./build-scripts/traceur/TraceurEs6ify");
+import ScssSass = require("./build-scripts/scss/ScssSass");
 
 
 function parseString(flagName: string, defaultValue: string = "") {
@@ -22,24 +23,23 @@ gulp.task("default", [], function () {
     var scenarioName = parseString("scenario");
     gutil.log("{ debug: " + debug + ", verboseCompileInfo: " + verboseCompileInfo + ", process.cwd: " + process.cwd() + " }");
 
+    var opts: AppPaths = {
+        entryFile: "./src/example1/SafePowerSource.js",
+        dstDir: "./build/",
+        dstFile: "app-babel-babelify.js",
+        srcPaths: ["node_modules", "./src/example1"],
+        projectRoot: process.cwd(),
+    };
+
     switch (scenarioName) {
         case "babel-babelify":
-            BabelBabelify.compileScripts(debug, verboseCompileInfo, {
-                entryFile: "./src/example1/SafePowerSource.js",
-                dstDir: "./build/",
-                dstFile: "app-babel-babelify.js",
-                srcPaths: ["node_modules", "./src/example1"],
-                projectRoot: process.cwd(),
-            });
+            BabelBabelify.compileScripts(debug, verboseCompileInfo, opts);
             return;
         case "traceur-es6ify":
-            TraceurEs6ify.compileScripts(debug, verboseCompileInfo, {
-                entryFile: "./src/example1/SafePowerSource.js",
-                dstDir: "./build/",
-                dstFile: "app-traceur-es6ify.js",
-                srcPaths: ["node_modules", "./src/example1"],
-                projectRoot: process.cwd(),
-            });
+            TraceurEs6ify.compileScripts(debug, verboseCompileInfo, opts);
+            return;
+        case "scss":
+            ScssSass.compileScripts(debug, verboseCompileInfo, opts);
             return;
         default:
             gutil.log("Error: unknown compile scenario '" + scenarioName + "'");