+++ CHANGELOG.md
@@ -2,8 +2,17 @@
 All notable changes to this project will be documented in this file.
 This project does its best to adhere to [Semantic Versioning](http://semver.org/).
 
+--------
+### [0.2.0](N/A) - 2016-06-17
+#### Added
+* Scss example compiling a .scss (aka SASS) file to .css using node-sass package
+
+#### Changed
+* Moved and renamed some functions
+* Expanded PathUtil with new checkFileAsciiEncodingSync(), objToString(), and exists*Sync() file and directory functions
+
 
 --------
-### [0.1.0](N/A) - 2016-06-16
+### [0.1.0](https://github.com/TeamworkGuy2/browser-bundle-examples/commit/dc20b31407021b5cdc588272a866192a25dff4a2) - 2016-06-16
 #### Added
 Initial commit of existing babel and traceur example build scripts, including BrowserifyHelper with helper functions for projects using browserify.